+++ src/renderer.ts
@@ -3,1246 +3,30 @@
  */
 
 import Column from './model/Column';
-import StringColumn from './model/StringColumn';
-import AnnotateColumn from './model/AnnotateColumn';
-import LinkColumn from './model/LinkColumn';
-import SelectionColumn from './model/SelectionColumn';
-import StackColumn from './model/StackColumn';
-import CategoricalColumn from './model/CategoricalColumn';
-import {INumberColumn} from './model/NumberColumn';
-import {forEach, attr, clipText, ITextRenderHints} from './utils';
-import {hsl} from 'd3';
-import {IDataRow} from './provider/ADataProvider';
-import * as d3 from 'd3';
-import MultiValueColumn from './model/MultiValueColumn';
-import SetColumn from './model/SetColumn';
-import {IBoxPlotColumn} from './model/BoxPlotColumn';
 import CategoricalNumberColumn from './model/CategoricalNumberColumn';
 import CompositeNumberColumn from './model/CompositeNumberColumn';
 
+import ICellRendererFactory from './renderers/ICellRendererFactory';
+import BarCellRenderer from './renderers/BarCellRenderer';
+import {DefaultCellRenderer} from './renderers/DefaultCellRenderer';
+import StringCellRenderer from './renderers/StringCellRenderer';
+import SelectionRenderer from './renderers/SelectionRenderer';
+import LinkCellRenderer from './renderers/LinkCellRenderer';
+import AnnotationRenderer from './renderers/AnnotationRenderer';
+import ActionRenderer from './renderers/ActionRenderer';
+import StackCellRenderer from './renderers/StackCellRenderer';
+import CategoricalCellRenderer from './renderers/CategoricalCellRenderer';
+import HeatmapCellRenderer from './renderers/HeatmapCellRenderer';
+import SparklineCellRenderer from './renderers/SparklineCellRenderer';
+import VerticalBarCellRenderer from './renderers/VerticalBarCellRenderer';
+import SetCellRenderer from './renderers/SetCellRenderer';
+import CircleCellRenderer from './renderers/CircleCellRenderer';
+import BoxplotCellRenderer from './renderers/BoxplotCellRenderer';
+import LoadingCellRenderer from './renderers/LoadingCellRenderer';
+import ThresholdCellRenderer from './renderers/ThresholdCellRenderer';
+import Heatmap from './renderers/Heatmap';
+import {IDOMRenderContext, ICanvasRenderContext} from './renderers/RendererContexts';
 
-/**
- * context for rendering, wrapped as an object for easy extensibility
- */
-export interface IRenderContext<T> {
-  /**
-   * the height of a row
-   * @param index
-   */
-  rowHeight(index: number): number;
-
-  /**
-   * render a column
-   * @param col
-   */
-  renderer(col: Column): T;
-
-  /**
-   * prefix used for all generated id names
-   */
-  readonly idPrefix: string;
-
-  /**
-   * lookup custom options by key
-   * @param key key to lookup
-   * @param default_ default value
-   */
-  option<T>(key: string, defaultValue: T): T;
-}
-
-export declare type IDOMRenderContext = IRenderContext<IDOMCellRenderer<Element>>;
-
-/**
- * a cell renderer for rendering a cell of specific column
- */
-export interface IDOMCellRenderer<T> {
-  /**
-   * template as a basis for the update
-   */
-  readonly template: string;
-  /**
-   * update a given node (create using the template) with the given data
-   * @param node the node to update
-   * @param d the data item
-   * @param i the order relative index
-   */
-  update(node: T, d: IDataRow, i: number): void;
-}
-export declare type ISVGCellRenderer = IDOMCellRenderer<SVGElement>;
-export declare type IHTMLCellRenderer = IDOMCellRenderer<HTMLElement>;
-
-
-export interface ICanvasRenderContext extends IRenderContext<CanvasRenderingContext2D> {
-  hovered(dataIndex: number): boolean;
-  selected(dataIndex: number): boolean;
-  readonly textHints: ITextRenderHints;
-}
-
-export interface ICanvasCellRenderer {
-  /**
-   * renders the current item
-   * @param ctx
-   * @param d
-   * @param i
-   */
-  (ctx: CanvasRenderingContext2D, d: IDataRow, i: number, dx: number, dy: number): void;
-}
-
-export interface ICellRendererFactory {
-  createSVG?(col: Column, context: IDOMRenderContext): ISVGCellRenderer;
-  createHTML?(col: Column, context: IDOMRenderContext): IHTMLCellRenderer;
-  createCanvas?(col: Column, context: ICanvasRenderContext): ICanvasCellRenderer;
-}
-/**
- * default renderer instance rendering the value as a text
- */
-export class DefaultCellRenderer implements ICellRendererFactory {
-  /**
-   * class to append to the text elements
-   * @type {string}
-   */
-  readonly textClass: string;
-  /**
-   * the text alignment: left, center, right
-   * @type {string}
-   */
-  readonly align: string;
-
-  constructor(textClass = 'text', align = 'left') {
-    this.textClass = textClass;
-    this.align = align;
-  }
-
-  createSVG(col: Column, context: IDOMRenderContext): ISVGCellRenderer {
-    return {
-      template: `<text class="${this.textClass}" clip-path="url(#cp${context.idPrefix}clipCol${col.id})"></text>`,
-      update: (n: SVGTextElement, d: IDataRow) => {
-        let alignmentShift = 2;
-        if (this.align === 'right') {
-          alignmentShift = col.getWidth() - 5;
-        } else if (this.align === 'center') {
-          alignmentShift = col.getWidth() * 0.5;
-        }
-        attr(n, {
-          x: alignmentShift
-        });
-        n.textContent = col.getLabel(d.v, d.dataIndex);
-      }
-    };
-  }
-
-  createHTML(col: Column, context: IDOMRenderContext): IHTMLCellRenderer {
-    return {
-      template: `<div class="${this.textClass} ${this.align}"></div>`,
-      update: (n: HTMLDivElement, d: IDataRow) => {
-        attr(n, {}, {
-          width: `${col.getWidth()}px`
-        });
-        n.textContent = col.getLabel(d.v, d.dataIndex);
-      }
-    };
-  }
-
-  createCanvas(col: Column, context: ICanvasRenderContext): ICanvasCellRenderer {
-    return (ctx: CanvasRenderingContext2D, d: IDataRow) => {
-      const bak = ctx.textAlign;
-      ctx.textAlign = this.align;
-      const w = col.getWidth();
-      let shift = 0;
-      if (this.align === 'center') {
-        shift = w / 2;
-      } else if (this.align === 'right') {
-        shift = w;
-      }
-      clipText(ctx, col.getLabel(d.v, d.dataIndex), shift, 0, w, context.textHints);
-      ctx.textAlign = bak;
-    };
-  }
-}
-
-
-class HeatmapCellRenderer implements ICellRendererFactory {
-
-  createSVG(col: MultiValueColumn, context: IDOMRenderContext): ISVGCellRenderer {
-    const cellDimension = col.calculateCellDimension(col.getWidth());
-    const colorScale = col.getColorScale();
-    const padding = context.option('rowBarPadding', 1);
-    return {
-
-      template: `<g class="heatmapcell"></g>`,
-      update: (n: SVGGElement, d: IDataRow, i: number) => {
-        const rect = d3.select(n).selectAll('rect').data(col.getValue(d.v, d.dataIndex));
-        rect.enter().append('rect').attr({
-          y: padding,
-          width: cellDimension,
-          height: context.rowHeight(i)
-        });
-        rect.attr({
-          x: (d, i) => i * cellDimension,
-          fill: colorScale
-        });
-        rect.exit().remove();
-      }
-    };
-  }
-
-
-  createHTML(col: MultiValueColumn, context: IDOMRenderContext): IHTMLCellRenderer {
-    const cellDimension = col.calculateCellDimension(col.getWidth());
-    const padding = context.option('rowBarPadding', 1);
-    const colorScale = col.getColorScale();
-
-    return {
-      template: `<div class="heatmapcell" style="top:${padding}px;"></div>`,
-      update: (n: HTMLDivElement, d: IDataRow, i: number) => {
-        const g = d3.select(n);
-        const div = g.selectAll('div').data(col.getValue(d.v, d.dataIndex));
-        div.enter().append('div').style({
-          widht: cellDimension + 'px',
-          height: context.rowHeight(i) + 'px'
-        });
-        div.style({
-          'background-color': colorScale,
-          'left': (d, i) => (i * cellDimension) + 'px'
-        });
-      }
-    };
-  }
-
-  createCanvas(col: MultiValueColumn, context: ICanvasRenderContext): ICanvasCellRenderer {
-    const cellDimension = col.calculateCellDimension(col.getWidth());
-    const padding = context.option('rowBarPadding', 1);
-    const colorScale = col.getColorScale();
-
-    return (ctx: CanvasRenderingContext2D, d: IDataRow, i: number) => {
-      const data = col.getValue(d.v, d.dataIndex);
-      const rowHeight = context.rowHeight(i);
-      data.forEach((d: number, j: number) => {
-        const x = (j * cellDimension);
-        ctx.beginPath();
-        ctx.fillStyle = colorScale(d);
-        ctx.fillRect(x, padding, cellDimension, rowHeight);
-      });
-    };
-  }
-
-}
-
-
-class SparklineCellRenderer implements ICellRendererFactory {
-
-  createSVG(col: MultiValueColumn, context: IDOMRenderContext): ISVGCellRenderer {
-    const scales = col.getSparklineScale();
-    const xScale = scales.xScale.range([0, col.getWidth()]);
-    const yScale = scales.yScale;
-    const line = d3.svg.line<number>()
-      .x((d, j) => xScale(j))
-      .y(yScale)
-      .interpolate('linear');
-    return {
-      template: `<path class="sparklinecell"></path>`,
-      update: (n: SVGGElement, d: IDataRow, i: number) => {
-        yScale.range([context.rowHeight(i), 0]);
-        attr(n, {
-          d: line(col.getValue(d.v, d.dataIndex))
-        });
-      }
-    };
-  }
-
-  createCanvas(col: MultiValueColumn, context: ICanvasRenderContext): ICanvasCellRenderer {
-    const scales = col.getSparklineScale();
-    const xScale = scales.xScale.range([0, col.getWidth()]);
-    const yScale = scales.yScale;
-
-    return (ctx: CanvasRenderingContext2D, d: IDataRow, i: number) => {
-      const data = col.getValue(d.v, d.dataIndex);
-      let xpos: number, ypos: number;
-      yScale.range([context.rowHeight(i), 0]);
-
-      ctx.strokeStyle = 'black';
-      ctx.fillStyle = 'black';
-      data.forEach((d, i) => {
-        ctx.beginPath();
-        ctx.moveTo(xpos, ypos);
-        xpos = xScale(i);
-        ypos = yScale(d);
-        ctx.lineTo(xpos, ypos);
-        ctx.stroke();
-        ctx.fill();
-      });
-    };
-  }
-}
-
-class ThresholdCellRenderer implements ICellRendererFactory {
-
-  createSVG(col: MultiValueColumn, context: IDOMRenderContext): ISVGCellRenderer {
-    const cellDimension = col.calculateCellDimension(col.getWidth());
-    const threshold = col.getThreshold();
-    const colorValues = col.getColorScale().range();
-
-    return {
-      template: `<g class="thresholdcell"></g>`,
-      update: (n: SVGGElement, d: IDataRow, i: number) => {
-        const rowHeight = context.rowHeight(i);
-        const rect = d3.select(n).selectAll('rect').data(col.getValue(d.v, d.dataIndex));
-        rect.enter().append('rect');
-        rect
-          .attr({
-            y: (d, j) => (d < threshold) ? (rowHeight / 2) : 0,
-            x: (d, j) => (j * cellDimension),
-            width: cellDimension,
-            height: (d, j) => (rowHeight / 2),
-            fill: (d) => (d < threshold) ? colorValues[0] : colorValues[2]
-          });
-        rect.exit().remove();
-      }
-    };
-  }
-
-  createCanvas(col: MultiValueColumn, context: ICanvasRenderContext): ICanvasCellRenderer {
-
-    const cellDimension = col.calculateCellDimension(col.getWidth());
-    const threshold = col.getThreshold();
-    const colorValues = col.getColorScale().range();
-
-    return (ctx: CanvasRenderingContext2D, d: IDataRow, i: number) => {
-      const data = col.getValue(d.v, d.dataIndex);
-      const rowHeight = context.rowHeight(i);
-      data.forEach((d, j) => {
-        ctx.beginPath();
-        const xpos = (j * cellDimension);
-        const ypos = (d < threshold) ? (rowHeight / 2) : 0;
-        ctx.fillStyle = (d < threshold) ? colorValues[0] : colorValues[2];
-        ctx.fillRect(xpos, ypos, cellDimension, rowHeight / 2);
-      });
-    };
-  }
-
-}
-
-
-function verticalBarScale(domain: number[], threshold: number, scale: d3.scale.Linear<number,number>, rowHeight: number) {
-  return (domain[0] < threshold) ? (scale.range([0, rowHeight / 2])) : scale.range([0, rowHeight]);
-}
-
-function verticalBarHeight(domain: number[], threshold: number, cellData: number, scale: d3.scale.Linear<number,number>, rowHeight: number) {
-  return (domain[0] < threshold) ? (rowHeight / 2 - scale(cellData)) : scale(cellData);
-
-}
-
-function verticalBarYpos(domain: number[], threshold: number, cellData: number, scale: d3.scale.Linear<number,number>, rowHeight: number) {
-  if (domain[0] < threshold) {
-    return (cellData < threshold) ? (rowHeight / 2) : rowHeight / 2 - scale(cellData);   // For positive and negative value
-  } else {
-    return rowHeight - scale(cellData);
-  }
-}
-
-class VerticalBarCellRenderer implements ICellRendererFactory {
-
-  createSVG(col: MultiValueColumn, context: IDOMRenderContext): ISVGCellRenderer {
-    const colorScale = col.getColorScale();
-    const cellDimension = col.calculateCellDimension(col.getWidth());
-    const defaultScale = col.getVerticalBarScale();
-    const threshold = col.getThreshold();
-    const domain = col.getDomain();
-
-    return {
-      template: `<g class="verticalbarcell"></g>`,
-      update: (n: SVGGElement, d: IDataRow, i: number) => {
-        const rowHeight = context.rowHeight(i);
-        const scale = verticalBarScale(domain, threshold, defaultScale, rowHeight);
-        const rect = d3.select(n).selectAll('rect').data(col.getValue(d.v, d.dataIndex));
-        rect.enter().append('rect').attr('width', cellDimension);
-        rect.attr({
-          y: (d) => verticalBarYpos(domain, threshold, d, scale, rowHeight),
-          x: (d, j) => (j * cellDimension),
-          height: (d: number) => verticalBarHeight(domain, threshold, d, scale, rowHeight),
-          fill: colorScale
-        });
-        rect.exit().remove();
-      }
-    };
-  }
-
-  createCanvas(col: MultiValueColumn, context: ICanvasRenderContext): ICanvasCellRenderer {
-
-    const colorScale = col.getColorScale();
-    const cellDimension = col.calculateCellDimension(col.getWidth());
-    const defaultScale = col.getVerticalBarScale();
-    const threshold = col.getThreshold();
-    const domain = col.getDomain();
-
-    return (ctx: CanvasRenderingContext2D, d: IDataRow, i: number) => {
-      const data = col.getValue(d.v, d.dataIndex);
-      const rowHeight = context.rowHeight(i);
-      const scale = verticalBarScale(domain, threshold, defaultScale, rowHeight);
-
-      data.forEach((d, j) => {
-        const xpos = (j * cellDimension);
-        const ypos = verticalBarYpos(domain, threshold, d, scale, rowHeight);
-        ctx.fillStyle = colorScale(d);
-        ctx.fillRect(xpos, ypos, cellDimension, verticalBarHeight(domain, threshold, d, scale, rowHeight));
-      });
-    };
-  }
-
-}
-
-class BoxplotCellRenderer implements ICellRendererFactory {
-
-  createSVG(col: IBoxPlotColumn & Column, context: IDOMRenderContext): ISVGCellRenderer {
-    const sortMethod = col.getSortMethod();
-    const topPadding = 2.5 * (context.option('rowBarPadding', 1));
-    const domain = col.getDomain();
-    const scale = d3.scale.linear().domain(domain).range([0, col.getWidth()]);
-    const sortedByMe = col.findMyRanker().getSortCriteria().col === col;
-    return {
-
-      template: `<g class="boxplotcell">
-            <rect class="boxplotrect" y="${topPadding}"></rect>
-            <path class="boxplotallpath"></path>
-            <path class="boxplotsortpath" style="display: none"></path>
-        </g>`,
-      update: (n: SVGGElement, d: IDataRow, i: number) => {
-        const rawBoxdata = col.getBoxPlotData(d.v, d.dataIndex);
-        const rowHeight = context.rowHeight(i);
-        const scaled = {
-          min: scale(rawBoxdata.min),
-          median: scale(rawBoxdata.median),
-          q1: scale(rawBoxdata.q1),
-          q3: scale(rawBoxdata.q3),
-          max: scale(rawBoxdata.max)
-        };
-        attr(n.querySelector('rect'), {
-          x: scaled.q1,
-          width: (scaled.q3 - scaled.q1),
-          height: (rowHeight - (topPadding * 2))
-        });
-        const bottomPos = (rowHeight - topPadding);
-        const middlePos = (rowHeight - topPadding) / 2;
-        const path = `M${scaled.min},${middlePos}L${scaled.q1},${middlePos}M${scaled.min},${topPadding}L${scaled.min},${bottomPos}` +   //minimum line
-          `M${scaled.median},${topPadding}L${scaled.median},${bottomPos}` +   //median line
-          `M${scaled.q3},${middlePos}L${scaled.max},${middlePos}` +
-          `M${scaled.max},${topPadding}L${scaled.max},${bottomPos}`;   // maximum line
-
-        attr(<SVGPathElement>n.querySelector('path.boxplotallpath'), {
-          d: path
-        });
-        attr(<SVGPathElement>n.querySelector('path.boxplotsortpath'), {
-          d: `M${scaled[sortMethod]},${topPadding}L${scaled[sortMethod]},${bottomPos}`
-        }, {
-          display: sortedByMe ? null : 'none'
-        });
-      }
-    };
-  }
-
-  createCanvas(col: IBoxPlotColumn & Column, context: ICanvasRenderContext): ICanvasCellRenderer {
-    const sortMethod = col.getSortMethod();
-    const topPadding = 2.5 * (context.option('rowBarPadding', 1));
-    const domain = col.getDomain();
-
-    const scale = d3.scale.linear().domain([d3.min(domain), d3.max(domain)]).range([0, col.getWidth()]);
-    const sortedByMe = col.findMyRanker().getSortCriteria().col === col;
-
-    return (ctx: CanvasRenderingContext2D, d: IDataRow, i: number) => {
-      const rowHeight = context.rowHeight(i);
-
-      // Rectangle
-      const rawBoxdata = col.getBoxPlotData(d.v, d.dataIndex);
-      const scaled = {
-        min: scale(rawBoxdata.min),
-        median: scale(rawBoxdata.median),
-        q1: scale(rawBoxdata.q1),
-        q3: scale(rawBoxdata.q3),
-        max: scale(rawBoxdata.max)
-      };
-      const minPos = scaled.min, maxPos = scaled.max, medianPos = scaled.median, q3Pos = scaled.q3, q1Pos = scaled.q1;
-      ctx.fillStyle = '#e0e0e0';
-      ctx.strokeStyle = 'black';
-      ctx.beginPath();
-      ctx.rect((q1Pos), topPadding, ((q3Pos) - (q1Pos)), (rowHeight - (topPadding * 2)));
-      ctx.fill();
-      ctx.stroke();
-
-      //Line
-      const bottomPos = (rowHeight - topPadding);
-      const middlePos = (rowHeight - topPadding) / 2;
-
-      ctx.strokeStyle = 'black';
-      ctx.fillStyle = '#e0e0e0';
-      ctx.beginPath();
-      ctx.moveTo(minPos, middlePos);
-      ctx.lineTo((q1Pos), middlePos);
-      ctx.moveTo(minPos, topPadding);
-      ctx.lineTo(minPos, bottomPos);
-      ctx.moveTo(medianPos, topPadding);
-      ctx.lineTo(medianPos, bottomPos);
-      ctx.moveTo((q3Pos), middlePos);
-      ctx.lineTo(maxPos, middlePos);
-      ctx.moveTo(maxPos, topPadding);
-      ctx.lineTo(maxPos, bottomPos);
-      ctx.stroke();
-      ctx.fill();
-
-
-      if (sortedByMe) {
-        ctx.strokeStyle = 'red';
-        ctx.fillStyle = '#ff0700';
-        ctx.beginPath();
-        ctx.moveTo(scaled[sortMethod], topPadding);
-        ctx.lineTo(scaled[sortMethod], bottomPos);
-        ctx.stroke();
-        ctx.fill();
-      }
-
-    };
-  }
-}
-
-function setPathCalculate(setData: boolean[], cellDimension: number) {
-
-  const catindexes = [];
-  setData.forEach((d: boolean, i: number) => (d) ? catindexes.push(i) : -1);
-
-  const left = (catindexes[0] * cellDimension) + (cellDimension / 2);
-  const right = (catindexes[catindexes.length - 1] * cellDimension) + (cellDimension / 2);
-
-  return {left, right};
-}
-
-
-class SetCellRenderer implements ICellRendererFactory {
-
-  createSVG(col: SetColumn, context: IDOMRenderContext): ISVGCellRenderer {
-    const cellDimension = col.cellDimension();
-    return {
-      template: `<g class="upsetcell"><path class="upsetpath"></path></g>`,
-      update: (n: SVGGElement, d: IDataRow, i: number) => {
-        const rowHeight = context.rowHeight(i);
-        const value = col.getBinaryValue(d.v, d.dataIndex);
-        const hasTrueValues = value.some((d) => d); //some values are true?
-
-        const circle = d3.select(n).selectAll('circle').data(value);
-        circle.enter().append('circle');
-        circle
-          .attr({
-            cy: (d, j) => (rowHeight / 2),
-            cx: (d, j) => (j * cellDimension) + (cellDimension / 2),
-            r: (cellDimension / 4),
-            class: (d) => d ? 'setcircle' : 'setcircleOpacity'
-          });
-        circle.exit().remove();
-
-        let path = '';
-        if (hasTrueValues) {
-          const pathCordinate = setPathCalculate(value, cellDimension);
-          path = `M${pathCordinate.left},${rowHeight / 2}L${pathCordinate.right},${rowHeight / 2}`;
-        }
-        attr(n.querySelector('path'), {
-          d: path
-        });
-      }
-    };
-  }
-
-  createCanvas(col: SetColumn, context: ICanvasRenderContext): ICanvasCellRenderer {
-
-    const cellDimension = col.cellDimension();
-    return (ctx: CanvasRenderingContext2D, d: IDataRow, i: number) => {
-      // Circle
-      const data = col.getBinaryValue(d.v, d.dataIndex);
-      const hasTrueValues = data.some((d) => d); //some values are true?
-      const rowHeight = context.rowHeight(i);
-      const radius = (rowHeight / 3);
-
-      ctx.save();
-      ctx.fillStyle = 'black';
-      ctx.strokeStyle = 'black';
-      if (hasTrueValues) {
-        const pathCordinate = hasTrueValues ? setPathCalculate(data, cellDimension) : null;
-        ctx.beginPath();
-        ctx.moveTo((pathCordinate.left), (rowHeight / 2));
-        ctx.lineTo((pathCordinate.right), (rowHeight / 2));
-        ctx.fill();
-        ctx.stroke();
-      }
-
-      data.forEach((d: boolean, j: number) => {
-        const posy = (rowHeight / 2);
-        const posx = (j * cellDimension) + (cellDimension / 2);
-        ctx.beginPath();
-        ctx.globalAlpha = d ? 1 : 0.1;
-        ctx.arc(posx, posy, radius, 0, 2 * Math.PI);
-        ctx.fill();
-        ctx.stroke();
-      });
-
-      ctx.restore();
-    };
-  }
-
-}
-
-class CircleCellRenderer implements ICellRendererFactory {
-  private readonly renderValue;
-
-  constructor(renderValue = false, private colorOf: (d: any, i: number, col: Column) => string = (d, i, col) => col.color) {
-    this.renderValue = renderValue;
-  }
-
-  createSVG(col: INumberColumn & Column, context: IDOMRenderContext): ISVGCellRenderer {
-    const padding = context.option('rowBarPadding', 1);
-    return {
-      template: `<g class="bar">
-          <circle class="${col.cssClass}" style="fill: ${col.color}">
-            <title></title>
-          </circle>
-          <text class="number ${this.renderValue ? '' : 'hoverOnly'}" clip-path="url(#cp${context.idPrefix}clipCol${col.id})"></text>
-        </g>`,
-      update: (n: SVGElement, d: IDataRow, i: number) => {
-        const v = col.getValue(d.v, d.dataIndex);
-        attr(<SVGCircleElement>n.querySelector('circle'), {
-          cy: (context.rowHeight(i) / 2),
-          cx: (col.getWidth() / 2),
-          r: (context.rowHeight(i) / 2) * v
-        });
-        attr(<SVGTextElement>n.querySelector('text'), {}).textContent = v;
-      }
-    };
-  }
-
-
-  createCanvas(col: INumberColumn & Column, context: ICanvasRenderContext): ICanvasCellRenderer {
-    return (ctx: CanvasRenderingContext2D, d: IDataRow, i: number) => {
-      const posy = (context.rowHeight(i) / 2);
-      const posx = (col.getWidth() / 2);
-      ctx.fillStyle = this.colorOf(d.v, i, col);
-      ctx.strokeStyle = this.colorOf(d.v, i, col);
-      ctx.beginPath();
-      ctx.arc(posx, posy, (context.rowHeight(i) / 2) * col.getValue(d.v, d.dataIndex), 0, 2 * Math.PI);
-      ctx.fill();
-      ctx.stroke();
-      if (this.renderValue || context.hovered(d.dataIndex) || context.selected(d.dataIndex)) {
-        ctx.fillStyle = context.option('style.text', 'black');
-        clipText(ctx, col.getLabel(d.v, d.dataIndex), 1, 0, col.getWidth() - 1, context.textHints);
-      }
-    };
-  }
-}
-
-
-/**
- * a renderer rendering a bar for numerical columns
- */
-export class BarCellRenderer implements ICellRendererFactory {
-  /**
-   * flag to always render the value
-   * @type {boolean}
-   */
-  private readonly renderValue;
-
-  constructor(renderValue = false, private colorOf: (d: any, i: number, col: Column) => string = (d, i, col) => col.color) {
-    this.renderValue = renderValue;
-  }
-
-  createSVG(col: INumberColumn & Column, context: IDOMRenderContext): ISVGCellRenderer {
-    const paddingTop = context.option('rowBarTopPadding', context.option('rowBarPadding', 1));
-    const paddingBottom = context.option('rowBarBottomPadding', context.option('rowBarPadding', 1));
-    return {
-      template: `<g class="bar">
-          <rect class="${col.cssClass}" y="${paddingTop}" style="fill: ${col.color}">
-            <title></title>
-          </rect>
-          <text class="number ${this.renderValue ? '' : 'hoverOnly'}" clip-path="url(#cp${context.idPrefix}clipCol${col.id})"></text>
-        </g>`,
-      update: (n: SVGGElement, d: IDataRow, i: number) => {
-        n.querySelector('rect title').textContent = col.getLabel(d.v, d.dataIndex);
-        const width = col.getWidth() * col.getValue(d.v, d.dataIndex);
-
-        attr(<SVGRectElement>n.querySelector('rect'), {
-          y: paddingTop,
-          width: isNaN(width) ? 0 : width,
-          height: context.rowHeight(i) - (paddingTop + paddingBottom)
-        }, {
-          fill: this.colorOf(d.v, i, col)
-        });
-        attr(<SVGTextElement>n.querySelector('text'), {}).textContent = col.getLabel(d.v, d.dataIndex);
-      }
-    };
-  }
-
-  createHTML(col: INumberColumn & Column, context: IDOMRenderContext): IHTMLCellRenderer {
-    const paddingTop = context.option('rowBarTopPadding', context.option('rowBarPadding', 1));
-    const paddingBottom = context.option('rowBarBottomPadding', context.option('rowBarPadding', 1));
-    return {
-      template: `<div class="bar" style="top:${paddingTop}px; background-color: ${col.color}">
-          <span class="number ${this.renderValue ? '' : 'hoverOnly'}"></span>
-        </div>`,
-      update: (n: HTMLDivElement, d: IDataRow, i: number) => {
-        const width = col.getWidth() * col.getValue(d.v, d.dataIndex);
-        attr(n, {
-          title: col.getLabel(d.v, d.dataIndex)
-        }, {
-          width: `${isNaN(width) ? 0 : width}px`,
-          height: `${context.rowHeight(i) - (paddingTop + paddingBottom)}px`,
-          top: `${paddingTop}px`,
-          'background-color': this.colorOf(d.v, i, col)
-        });
-        n.querySelector('span').textContent = col.getLabel(d.v, d.dataIndex);
-      }
-    };
-  }
-
-  createCanvas(col: INumberColumn & Column, context: ICanvasRenderContext): ICanvasCellRenderer {
-    const paddingTop = context.option('rowBarTopPadding', context.option('rowBarPadding', 1));
-    const paddingBottom = context.option('rowBarBottomPadding', context.option('rowBarPadding', 1));
-    return (ctx: CanvasRenderingContext2D, d: IDataRow, i: number) => {
-      ctx.fillStyle = this.colorOf(d.v, i, col);
-      const width = col.getWidth() * col.getValue(d.v, d.dataIndex);
-      ctx.fillRect(0, paddingTop, isNaN(width) ? 0 : width, context.rowHeight(i) - (paddingTop + paddingBottom));
-      if (this.renderValue || context.hovered(d.dataIndex) || context.selected(d.dataIndex)) {
-        ctx.fillStyle = context.option('style.text', 'black');
-        clipText(ctx, col.getLabel(d.v, d.dataIndex), 1, 0, col.getWidth() - 1, context.textHints);
-      }
-    };
-  }
-}
-
-function toHeatMapColor(d: any, index: number, col: INumberColumn & Column) {
-  let v = col.getNumber(d, index);
-  if (isNaN(v)) {
-    v = 0;
-  }
-  //hsl space encoding, encode in lightness
-  const color = hsl(col.color || Column.DEFAULT_COLOR);
-  color.l = v;
-  return color.toString();
-}
-
-const heatmap = {
-  createSVG(col: INumberColumn & Column, context: IDOMRenderContext): ISVGCellRenderer {
-    const padding = context.option('rowBarPadding', 1);
-    return {
-      template: `<rect class="heatmap ${col.cssClass}" y="${padding}" style="fill: ${col.color}">
-            <title></title>
-          </rect>`,
-      update: (n: SVGGElement, d: IDataRow, i: number) => {
-        n.querySelector('title').textContent = col.getLabel(d.v, d.dataIndex);
-        const w = context.rowHeight(i) - padding * 2;
-
-        attr(n, {
-          y: padding,
-          width: w,
-          height: w
-        }, {
-          fill: toHeatMapColor(d.v, d.dataIndex, col)
-        });
-      }
-    };
-  },
-  createHTML(col: INumberColumn & Column, context: IDOMRenderContext): IHTMLCellRenderer {
-    const padding = context.option('rowBarPadding', 1);
-    return {
-      template: `<div class="heatmap ${col.cssClass}" style="background-color: ${col.color}; top: ${padding}"></div>`,
-      update: (n: HTMLElement, d: IDataRow, i: number) => {
-        const w = context.rowHeight(i) - padding * 2;
-        attr(n, {
-          title: col.getLabel(d.v, d.dataIndex)
-        }, {
-          width: `${w}px`,
-          height: `${w}px`,
-          top: `${padding}px`,
-          'background-color': toHeatMapColor(d.v, d.dataIndex, col)
-        });
-      }
-    };
-  },
-  createCanvas(col: INumberColumn & Column, context: ICanvasRenderContext): ICanvasCellRenderer {
-    const padding = context.option('rowBarPadding', 1);
-    return (ctx: CanvasRenderingContext2D, d: IDataRow, i: number) => {
-      const w = context.rowHeight(i) - padding * 2;
-      ctx.fillStyle = toHeatMapColor(d.v, d.dataIndex, col);
-      ctx.fillRect(padding, padding, w, w);
-    };
-  }
-};
-
-const action = {
-  createSVG(col: Column, context: IDOMRenderContext): ISVGCellRenderer {
-    const actions = context.option('actions', []);
-    return {
-      template: `<text class="actions hoverOnly fa">${actions.map((a) => `<tspan>${a.icon}</tspan>`).join('')}</text>`,
-      update: (n: SVGTextElement, d: IDataRow) => {
-        forEach(n, 'tspan', (ni: SVGTSpanElement, i: number) => {
-          ni.onclick = function (event) {
-            event.preventDefault();
-            event.stopPropagation();
-            actions[i].action(d.v, d.dataIndex);
-          };
-        });
-      }
-    };
-  },
-  createHTML(col: Column, context: IDOMRenderContext): IHTMLCellRenderer {
-    const actions = context.option('actions', []);
-    return {
-      template: `<div class="actions hoverOnly">${actions.map((a) => `<span title="${a.name}" class="fa">${a.icon}</span>`).join('')}</div>`,
-      update: (n: HTMLElement, d: IDataRow) => {
-        forEach(n, 'span', (ni: HTMLSpanElement, i: number) => {
-          ni.onclick = function (event) {
-            event.preventDefault();
-            event.stopPropagation();
-            actions[i].action(d.v, d.dataIndex);
-          };
-        });
-      }
-    };
-  },
-  createCanvas(col: Column, context: ICanvasRenderContext): ICanvasCellRenderer {
-    const actions = context.option('actions', []);
-    return (ctx: CanvasRenderingContext2D, d: IDataRow, i: number, dx: number, dy: number) => {
-      const hovered = context.hovered(d.dataIndex);
-      if (hovered) {
-        const overlay = showOverlay(context.idPrefix + col.id, dx, dy);
-        overlay.style.width = col.getWidth() + 'px';
-        overlay.classList.add('actions');
-        overlay.innerHTML = actions.map((a) => `<span title="${a.name}" class="fa">${a.icon}</span>`).join('');
-        forEach(overlay, 'span', (ni: HTMLSpanElement, i) => {
-          ni.onclick = function (event) {
-            event.preventDefault();
-            event.stopPropagation();
-            actions[i].action(d.v, d.dataIndex);
-          };
-        });
-      }
-    };
-  }
-};
-
-const selection = {
-  createSVG(col: SelectionColumn): ISVGCellRenderer {
-    return {
-      template: `<text class="selection fa"><tspan class="selectionOnly">\uf046</tspan><tspan class="notSelectionOnly">\uf096</tspan></text>`,
-      update: (n: SVGGElement, d: IDataRow) => {
-        n.onclick = function (event) {
-          event.preventDefault();
-          event.stopPropagation();
-          col.toggleValue(d.v, d.dataIndex);
-        };
-      }
-    };
-  },
-  createHTML(col: SelectionColumn): IHTMLCellRenderer {
-    return {
-      template: `<div class="selection fa"></div>`,
-      update: (n: HTMLElement, d: IDataRow) => {
-        n.onclick = function (event) {
-          event.preventDefault();
-          event.stopPropagation();
-          col.toggleValue(d.v, d.dataIndex);
-        };
-      }
-    };
-  },
-  createCanvas(col: SelectionColumn, context: ICanvasRenderContext): ICanvasCellRenderer {
-    return (ctx: CanvasRenderingContext2D, d: IDataRow) => {
-      const bak = ctx.font;
-      ctx.font = '10pt FontAwesome';
-      clipText(ctx, col.getValue(d.v, d.dataIndex) ? '\uf046' : '\uf096', 0, 0, col.getWidth(), context.textHints);
-      ctx.font = bak;
-    };
-  }
-};
-
-const annotate = {
-  createSVG(col: AnnotateColumn, context: IDOMRenderContext): ISVGCellRenderer {
-    return {
-      template: `<g class="annotations">
-        <text class="notHoverOnly text" clip-path="url(#cp${context.idPrefix}clipCol${col.id})"></text>
-        <foreignObject class="hoverOnly" x="-2" y="-2">
-          <input type="text">
-        </foreignObject>
-       </g>`,
-      update: (n: SVGGElement, d: IDataRow, i: number) => {
-        const input: HTMLInputElement = <HTMLInputElement>n.querySelector('foreignObject *');
-        input.onchange = function () {
-          col.setValue(d.v, d.dataIndex, this.value);
-        };
-        input.onclick = function (event) {
-          event.stopPropagation();
-        };
-        input.style.width = col.getWidth() + 'px';
-        input.value = col.getLabel(d.v, d.dataIndex);
-
-        n.querySelector('text').textContent = col.getLabel(d.v, d.dataIndex);
-        const f = n.querySelector('foreignObject');
-        f.setAttribute('width', String(col.getWidth()));
-        f.setAttribute('height', String(context.rowHeight(i)));
-      }
-    };
-  },
-  createHTML(col: AnnotateColumn): IHTMLCellRenderer {
-    return {
-      template: `<div class="annotations text">
-        <input type="text" class="hoverOnly">
-        <span class="text notHoverOnly"></span>
-       </div>`,
-      update: (n: HTMLElement, d: IDataRow) => {
-        const input: HTMLInputElement = <HTMLInputElement>n.querySelector('input');
-        input.onchange = function () {
-          col.setValue(d.v, d.dataIndex, this.value);
-        };
-        input.onclick = function (event) {
-          event.stopPropagation();
-        };
-        n.style.width = input.style.width = col.getWidth() + 'px';
-        input.value = col.getLabel(d.v, d.dataIndex);
-        n.querySelector('span').textContent = col.getLabel(d.v, d.dataIndex);
-      }
-    };
-  },
-  createCanvas(col: AnnotateColumn, context: ICanvasRenderContext): ICanvasCellRenderer {
-    return (ctx: CanvasRenderingContext2D, d: IDataRow, i: number, dx: number, dy: number) => {
-      const hovered = context.hovered(d.dataIndex);
-      if (hovered) {
-        const overlay = showOverlay(context.idPrefix + col.id, dx, dy);
-        overlay.style.width = col.getWidth() + 'px';
-        overlay.innerHTML = `<input type="text" value="${col.getValue(d.v, d.dataIndex)}" style="width:${col.getWidth()}px">`;
-        const input = <HTMLInputElement>overlay.childNodes[0];
-        input.onchange = function () {
-          col.setValue(d.v, d.dataIndex, this.value);
-        };
-        input.onclick = function (event) {
-          event.stopPropagation();
-        };
-      } else {
-        clipText(ctx, col.getLabel(d.v, d.dataIndex), 0, 0, col.getWidth(), context.textHints);
-      }
-    };
-  }
-};
-
-function showOverlay(id: string, dx: number, dy: number) {
-  let overlay = <HTMLDivElement>document.querySelector(`div.lu-overlay#O${id}`);
-  if (!overlay) {
-    overlay = document.createElement('div');
-    overlay.classList.add('lu-overlay');
-    overlay.id = 'O' + id;
-    document.querySelector('.lu-body').appendChild(overlay);
-  }
-  overlay.style.display = 'block';
-  overlay.style.left = dx + 'px';
-  overlay.style.top = dy + 'px';
-  return overlay;
-}
-
-export function hideOverlays() {
-  forEach(document.querySelector('div.lu-body'), 'div.lu-overlay', (d: HTMLDivElement) => d.style.display = null);
-}
-
-const link = {
-  createSVG(col: LinkColumn, context: IDOMRenderContext): ISVGCellRenderer {
-    return {
-      template: `<text class="link text" clip-path="url(#cp${context.idPrefix}clipCol${col.id})"></text>`,
-      update: (n: SVGTextElement, d: IDataRow) => {
-        n.innerHTML = col.isLink(d.v, d.dataIndex) ? `<a class="link" xlink:href="${col.getValue(d.v, d.dataIndex)}" target="_blank">${col.getLabel(d.v, d.dataIndex)}</a>` : col.getLabel(d.v, d.dataIndex);
-      }
-    };
-  },
-  createHTML(col: LinkColumn): IHTMLCellRenderer {
-    return {
-      template: `<div class="link text"></div>`,
-      update: (n: HTMLElement, d: IDataRow) => {
-        n.style.width = col.getWidth() + 'px';
-        n.innerHTML = col.isLink(d.v, d.dataIndex) ? `<a class="link" href="${col.getValue(d.v, d.dataIndex)}" target="_blank">${col.getLabel(d.v, d.dataIndex)}</a>` : col.getLabel(d.v, d.dataIndex);
-      }
-    };
-  },
-  createCanvas(col: LinkColumn, context: ICanvasRenderContext): ICanvasCellRenderer {
-    return (ctx: CanvasRenderingContext2D, d: IDataRow, i: number, dx: number, dy: number) => {
-      const isLink = col.isLink(d.v, d.dataIndex);
-      if (!isLink) {
-        clipText(ctx, col.getLabel(d.v, d.dataIndex), 0, 0, col.getWidth(), context.textHints);
-        return;
-      }
-      const hovered = context.hovered(d.dataIndex);
-      if (hovered) {
-        const overlay = showOverlay(context.idPrefix + col.id, dx, dy);
-        overlay.style.width = col.getWidth() + 'px';
-        overlay.innerHTML = `<a class="link" href="${col.getValue(d.v, d.dataIndex)}" target="_blank">${col.getLabel(d.v, d.dataIndex)}</a>`;
-      } else {
-        const bak = ctx.fillStyle;
-        ctx.fillStyle = context.option('style.link', context.option('style.text', 'black'));
-        clipText(ctx, col.getLabel(d.v, d.dataIndex), 0, 0, col.getWidth(), context.textHints);
-        ctx.fillStyle = bak;
-      }
-    };
-  }
-};
-
-/**
- * renders a string with additional alignment behavior
- * one instance factory shared among strings
- */
-class StringCellRenderer {
-  private alignments = {
-    left: new DefaultCellRenderer(),
-    right: new DefaultCellRenderer('text_right', 'right'),
-    center: new DefaultCellRenderer('text_center', 'center')
-  };
-
-  createSVG(col: StringColumn, context: IDOMRenderContext): ISVGCellRenderer {
-    return this.alignments[col.alignment].createSVG(col, context);
-  }
-
-  createHTML(col: StringColumn, context: IDOMRenderContext): IHTMLCellRenderer {
-    return this.alignments[col.alignment].createHTML(col, context);
-  }
-
-  createCanvas(col: StringColumn, context: ICanvasRenderContext): ICanvasCellRenderer {
-    return this.alignments[col.alignment].createCanvas(col, context);
-  }
-}
-
-/**
- * renders categorical columns as a colored rect with label
- */
-export class CategoricalCellRenderer implements ICellRendererFactory {
-  /**
-   * class to append to the text elements
-   * @type {string}
-   */
-  readonly textClass;
-
-  constructor(textClass = 'cat') {
-    this.textClass = textClass;
-  }
-
-  createSVG(col: CategoricalColumn, context: IDOMRenderContext): ISVGCellRenderer {
-    const padding = context.option('rowBarPadding', 1);
-    return {
-      template: `<g class="${this.textClass}">
-        <text clip-path="url(#cp${context.idPrefix}clipCol${col.id})"></text>
-        <rect y="${padding}"></rect>
-      </g>`,
-      update: (n: SVGGElement, d: IDataRow, i: number) => {
-        const cell = Math.min(col.getWidth() * 0.3, Math.max(context.rowHeight(i) - padding * 2, 0));
-
-        attr(<SVGRectElement>n.querySelector('rect'), {
-          width: cell,
-          height: cell
-        }, {
-          fill: col.getColor(d.v, d.dataIndex)
-        });
-        attr(<SVGTextElement>n.querySelector('text'), {
-          x: cell + padding * 2
-        }).textContent = col.getLabel(d.v, d.dataIndex);
-      }
-    };
-  }
-
-  createHTML(col: CategoricalColumn, context: IDOMRenderContext): IHTMLCellRenderer {
-    const padding = context.option('rowBarPadding', 1);
-    return {
-      template: `<div class="${this.textClass}">
-        <div></div>
-        <span></span>
-      </div>`,
-      update: (n: HTMLElement, d: IDataRow, i: number) => {
-        const cell = Math.min(col.getWidth() * 0.3, Math.max(context.rowHeight(i) - padding * 2, 0));
-        attr(n, {}, {
-          width: `${col.getWidth()}px`
-        });
-        attr(<HTMLDivElement>n.querySelector('div'), {}, {
-          width: cell + 'px',
-          height: cell + 'px',
-          'background-color': col.getColor(d.v, d.dataIndex)
-        });
-        attr(<HTMLSpanElement>n.querySelector('span'), {}).textContent = col.getLabel(d.v, d.dataIndex);
-      }
-    };
-  }
-
-  createCanvas(col: CategoricalColumn, context: ICanvasRenderContext): ICanvasCellRenderer {
-    const padding = context.option('rowBarPadding', 1);
-    return (ctx: CanvasRenderingContext2D, d: IDataRow, i: number) => {
-      const cell = Math.min(col.getWidth() * 0.3, Math.max(context.rowHeight(i) - padding * 2, 0));
-      ctx.fillStyle = col.getColor(d.v, d.dataIndex);
-      ctx.fillRect(0, 0, cell, cell);
-      ctx.fillStyle = context.option('style.text', 'black');
-      clipText(ctx, col.getLabel(d.v, d.dataIndex), cell + 2, 0, col.getWidth() - cell - 2, context.textHints);
-    };
-  };
-}
-
-/**
- * machtes the columns and the dom nodes representing them
- * @param node
- * @param columns
- * @param helperType
- */
-export function matchColumns(node: SVGGElement | HTMLElement, columns: {column: Column, renderer: IDOMCellRenderer<any>}[], helperType = 'svg') {
-  if (node.childElementCount === 0) {
-    // initial call fast method
-    node.innerHTML = columns.map((c) => c.renderer.template).join('');
-    columns.forEach((col, i) => {
-      const cnode = <Element>node.childNodes[i];
-      // set attribute for finding again
-      cnode.setAttribute('data-column-id', col.column.id);
-      // store current renderer
-      cnode.setAttribute('data-renderer', col.column.getRendererType());
-    });
-    return;
-  }
-
-  function matches(c: {column: Column}, i: number) {
-    //do both match?
-    const n = <Element>(node.childElementCount <= i ? null : node.childNodes[i]);
-    return n != null && n.getAttribute('data-column-id') === c.column.id && n.getAttribute('data-renderer') === c.column.getRendererType();
-  }
-
-  if (columns.every(matches)) {
-    return; //nothing to do
-  }
-
-  const idsAndRenderer = new Set(columns.map((c) => c.column.id + '@' + c.column.getRendererType()));
-  //remove all that are not existing anymore
-  Array.prototype.slice.call(node.childNodes).forEach((n) => {
-    const id = n.getAttribute('data-column-id');
-    const renderer = n.getAttribute('data-renderer');
-    const idAndRenderer = id + '@' + renderer;
-    if (!idsAndRenderer.has(idAndRenderer)) {
-      node.removeChild(n);
-    }
-  });
-  const helper = helperType === 'svg' ? document.createElementNS('http://www.w3.org/2000/svg', 'g') : document.createElement('div');
-  columns.forEach((col) => {
-    let cnode = node.querySelector(`[data-column-id="${col.column.id}"]`);
-    if (!cnode) {
-      //create one
-      helper.innerHTML = col.renderer.template;
-      cnode = <Element>helper.childNodes[0];
-      cnode.setAttribute('data-column-id', col.column.id);
-      cnode.setAttribute('data-renderer', col.column.getRendererType());
-    }
-    node.appendChild(cnode);
-  });
-}
-
-/**
- * renders a stacked column using composite pattern
- */
-class StackCellRenderer implements ICellRendererFactory {
-  constructor(private readonly nestingPossible: boolean = true) {
-  }
-
-  private createData(col: StackColumn, context: IRenderContext<any>) {
-    const stacked = this.nestingPossible && context.option('stacked', true);
-    const padding = context.option('columnPadding', 0);
-    let offset = 0;
-    return col.children.map((d) => {
-      const shift = offset;
-      offset += d.getWidth();
-      offset += (!stacked ? padding : 0);
-      return {
-        column: d,
-        shift,
-        stacked,
-        renderer: context.renderer(d)
-      };
-    });
-  }
-
-  createSVG(col: StackColumn, context: IDOMRenderContext): ISVGCellRenderer {
-    const cols = this.createData(col, context);
-    return {
-      template: `<g class="stack component${context.option('stackLevel', 0)}">${cols.map((d) => d.renderer.template).join('')}</g>`,
-      update: (n: SVGGElement, d: IDataRow, i: number) => {
-        let stackShift = 0;
-        matchColumns(n, cols);
-        cols.forEach((col, ci) => {
-          const cnode: any = n.childNodes[ci];
-          cnode.setAttribute('transform', `translate(${col.shift - stackShift},0)`);
-          col.renderer.update(cnode, d, i);
-          if (col.stacked) {
-            stackShift += col.column.getWidth() * (1 - col.column.getValue(d.v, d.dataIndex));
-          }
-        });
-      }
-    };
-  }
-
-  createHTML(col: StackColumn, context: IDOMRenderContext): IHTMLCellRenderer {
-    const cols = this.createData(col, context);
-    return {
-      template: `<div class="stack component${context.option('stackLevel', 0)}">${cols.map((d) => d.renderer.template).join('')}</div>`,
-      update: (n: HTMLDivElement, d: IDataRow, i: number) => {
-        let stackShift = 0;
-        matchColumns(n, cols, 'html');
-        cols.forEach((col, ci) => {
-          const cnode: any = n.childNodes[ci];
-          cnode.style.transform = `translate(${col.shift - stackShift}px,0)`;
-          col.renderer.update(cnode, d, i);
-          if (col.stacked) {
-            stackShift += col.column.getWidth() * (1 - col.column.getValue(d.v, d.dataIndex));
-          }
-        });
-      }
-    };
-  }
-
-  createCanvas(col: StackColumn, context: ICanvasRenderContext): ICanvasCellRenderer {
-    const cols = this.createData(col, context);
-    return (ctx: CanvasRenderingContext2D, d: IDataRow, i: number, dx: number, dy: number) => {
-      let stackShift = 0;
-      cols.forEach((col) => {
-        const shift = col.shift - stackShift;
-        ctx.translate(shift, 0);
-        col.renderer(ctx, d, i, dx + shift, dy);
-        ctx.translate(-shift, 0);
-        if (col.stacked) {
-          stackShift += col.column.getWidth() * (1 - col.column.getValue(d.v, d.dataIndex));
-        }
-      });
-    };
-  }
-}
-
-
-const loading = {
-  createSVG(col: Column): ISVGCellRenderer {
-    return {
-      template: `<text class="loading"><tspan class="fa">\uf110</tspan>Loading…</text>`,
-      update: () => undefined // TODO svg animation
-    };
-  },
-  createHTML(col: Column): IHTMLCellRenderer {
-    return {
-      template: `<div class="loading"><i class="fa fa-spinner fa-pulse"></i><div>Loading…</div></div>`,
-      update: () => undefined
-    };
-  },
-  createCanvas(col: Column, context: ICanvasRenderContext): ICanvasCellRenderer {
-    const base = Date.now() % 360;
-    return (ctx: CanvasRenderingContext2D, d: IDataRow, i: number) => {
-      clipText(ctx, 'Loading…', 10, 0, col.getWidth() - 10, context.textHints);
-      const angle = (base + i * 45) * (Math.PI / 180);
-      ctx.save();
-      ctx.font = '10pt FontAwesome';
-      ctx.textAlign = 'center';
-      const shift = (context.rowHeight(i) - context.textHints.spinnerWidth) * 0.5;
-      ctx.translate(2, shift + context.textHints.spinnerWidth * 0.5);
-      ctx.rotate(angle);
-      ctx.translate(0, -context.textHints.spinnerWidth * 0.5);
-      ctx.fillText('\uf110', 0, 0);
-      ctx.restore();
-    };
-  }
-};
 
 export const defaultCellRenderer = new DefaultCellRenderer();
 const combineCellRenderer = new BarCellRenderer(false, (d, i, col: CompositeNumberColumn) => col.getColor(d, i));
@@ -1256,11 +40,11 @@ export const renderers: {[key: string]: ICellRendererFactory} = {
   number: new BarCellRenderer(),
   ordinal: new BarCellRenderer(true, (d, i, col: CategoricalNumberColumn) => col.getColor(d, i)),
   string: new StringCellRenderer(),
-  selection,
-  heatmap,
-  link,
-  annotate,
-  actions: action,
+  selection: new SelectionRenderer(),
+  heatmap: new Heatmap(),
+  link: new LinkCellRenderer(),
+  annotate: new AnnotationRenderer(),
+  actions: new ActionRenderer(),
   stack: new StackCellRenderer(),
   nested: new StackCellRenderer(false),
   categorical: new CategoricalCellRenderer(),
@@ -1275,7 +59,7 @@ export const renderers: {[key: string]: ICellRendererFactory} = {
   set: new SetCellRenderer(),
   circle: new CircleCellRenderer(),
   boxplot: new BoxplotCellRenderer(),
-  loading
+  loading: new LoadingCellRenderer()
 };
 
 function chooseRenderer(col: Column, renderers: {[key: string]: ICellRendererFactory}): ICellRendererFactory {
